feat(param): パラメータをBIGINT化してレガシー準拠の青天井にする
レガシーではスキルパラメータに上限がなかった(青天井)。rewrite側で暫定的に
設けていた999上限を撤廃し、INT(約21億)の型上限も超えられるようBIGINTへ拡張する。

- migration 0036: 16スキルパラメータ + energy/energy_max/nou_energy/nou_energy_max をBIGINT化
  (satiety/weight_g/disease_index/height_cm は独自の小範囲なのでINTのまま)
- RefreshPowerMax の派生パワー上限計算を ::int → ::bigint 化(巨大パラメータでのオーバーフロー回避)
- detailedParamMax を 999 → 1e15(int64/BIGINT上限より十分小さいオーバーフローガード。ゲーム上の上限ではない)

+++ app/backend/internal/player/player.go
@@ -195,9 +195,9 @@ func RefreshPowerMax(ctx context.Context, tx pgx.Tx, playerID int64) error {
 		FROM (
 			SELECT player_id,
 				(FLOOR(looks/12.0 + tairyoku/4.0 + kenkou/4.0 + speed/8.0
-				       + power/8.0 + wanryoku/8.0 + kyakuryoku/8.0) + 1)::int AS emax,
+				       + power/8.0 + wanryoku/8.0 + kyakuryoku/8.0) + 1)::bigint AS emax,
 				(FLOOR(kokugo/6.0 + suugaku/6.0 + rika/6.0 + syakai/6.0
-				       + eigo/6.0 + ongaku/6.0 + bijutsu/6.0) + 1)::int AS nmax
+				       + eigo/6.0 + ongaku/6.0 + bijutsu/6.0) + 1)::bigint AS nmax
 			FROM player_status WHERE player_id = $1
 		) m
 		WHERE ps.player_id = m.player_id`, playerID); err != nil {
